rename internal ref: IfWrapper -> hal
