+++ Applications/utility_command.c
@@ -1,7 +1,7 @@
 #pragma section REPRO
 #include "utility_command.h"
 #include "../System/ApplicationManager/app_info.h"
-#include "../IfWrapper/uart.h"
+#include "../hal/uart.h"
 #include <src_user/Settings/port_config.h>
 #include <string.h>                     // for memcpy
 #include "../TlmCmd/common_cmd_packet_util.h"
